modify: move resource control logic of avatar provider
